requirements.txt
scipy==1.3.1
tensorflow==1.14.0
numpy==1.16.4
Flask_SQLAlchemy==2.4.1
six==1.12.0
imutils==0.5.3
Flask==1.1.1
Pillow==6.2.0
imageio==2.6.1
scikit_learn==0.21.3
